==> step1 generate_train_daga.ipynb <==
import os
import random
import json
import openai
import requests
import wikipediaapi
import itables
import numpy as np
import pandas as pd
import plotly.express as px
from datasets import load_dataset
from pathlib import Path
from tqdm import tqdm

# specify OpenAI API key in Kaggle's secrets add-ons.


# openai.api_base = "https://www.plus6.plus/vl"
openai.api_key = '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'

# https://www.kaggle.com/code/nbroad/create-science-wikipedia-dataset
files = list(map(str, Path("data/wiki-20220301-en-sci").glob("*.parquet")))
ds = load_dataset("parquet", data_files=files, split="train")

text = ds[0]['text']

import requests
import json

url = "https://api.openai.com/v1/chat/completions"

payload = json.dumps({
  "model": "gpt-3.5-turbo",
  "temperature": 0.9,
  "max_tokens": 500,
  "messages": [
    {
        'role':'system', 
        'content': system_message
    },
    {
      "role": "user",
      "content": text
    }
  ],
  "stream": False
})
headers = {
  'Content-Type': 'application/json',
  'Authorization': 'Bearer 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'
}

response = requests.request("POST", url, headers=headers, data=payload)

print(response.text)

options_set = set(("option_1", "option_2", "option_3", "option_4", "option_5"))
response_keys_set = set(("question", "option_1", "option_2", "option_3", "option_4", "option_5", "answer"))

delimiter = "####"
system_message = f"""
You will be provided with TEXT from wikipedia. \
The TEXT will be delimited with {delimiter} characters.
Output a python list of 5 dict objects, where each object is \
a multiple choice question whom answers should be in \
the given TEXT and that has 5 choices each and has the following format:
    'question': <question on the TEXT>
    'option_1': <question answer option>
    'option_2': <question answer option>
    'option_3': <question answer option>
    'option_4': <question answer option>
    'option_5': <question answer option>
    'answer': <answer option key label>

You should tell me which one of your proposed options is right \
by assigning the corresponding option's key label in the 'answer' field.

The question, the answer and question answer options should be broad, \
challenging, long, detailed and based on the TEXT provided.

Only output the list of objects, with nothing else.
"""

def get_completion_messages(wiki_text):
    return [  
        {
            'role':'system', 
            'content': system_message
        },    
        {
            'role':'user', 
            'content': f"{delimiter}{wiki_text}{delimiter}"
        },  
    ]

def get_completion_from_messages(
    messages, 
    model="gpt-3.5-turbo-0613", 
    temperature=0.8, 
    max_tokens=3500
):
    try:

        response = openai.ChatCompletion.create(
            model= model,
            messages=messages,
            temperature = temperature,
            max_tokens = max_tokens,
            stream=True,

        )
        
        completion = {'role': '', 'content': ''}

        for event in response:

            if event['choices'][0]['finish_reason'] == 'stop':

                #print(f'收到的完成数据: {completion}')

                break

            for delta_k, delta_v in event['choices'][0]['delta'].items():

                #print(f'流响应数据: {delta_k} = {delta_v}')

                completion[delta_k] += delta_v

        messages.append(completion)  # 直接在传入参数 messages 中追加消息

        return (True, '')

    except Exception as err:

        return (False, f'OpenAI API 异常: {err}')

def is_correctly_formatted(mcq) -> bool:
    """检查生成的数据是否规范"""
    return all([
        len(el) == len(response_keys_set) and response_keys_set == set(list(el.keys()))
        for el in mcq
    ])


max_completion_attempts = 5
multiple_choice_questions = []
count = 0
thread = 10000
for index, row in tqdm(enumerate(ds), total=len(ds)):
   
    text = row['text'][:3500]
    message = get_completion_messages(text)    
    if count >= thread:
        break
    attempts_counter = 0
    while True:
        try:
            get_completion_from_messages(message)
            assert len(message) == 3
            chatgpt_response = message[-1]['content']
            #print(chatgpt_response)
            mcq = eval(chatgpt_response)
            # 每个prompt生成5条数据
            if not isinstance(mcq, list) or len(mcq) < 5 or not is_correctly_formatted(mcq):
                raise Exception

            for i in range(len(mcq)): # 遍历每条数据
                mcq[i]["ori_index"] = index # 添加来源prompt
                mcq[i]["ori_wiki"] = text # 添加参考wiki文章
                if mcq[i]["answer"] in options_set:
                    continue
                else:
                    # index method will raise an error if answer isn't in list
                    answ_indx = [v.lower() for v in mcq[i].values()].index(mcq[i]["answer"].lower())
                    mcq[i]["answer"] = list(mcq[i].keys())[answ_indx]

            multiple_choice_questions += mcq
            #print("Generated count:", index+1)
            break
        except Exception:
            count += 1
            attempts_counter += 1
            print("Attempts count:", attempts_counter)
            if attempts_counter > max_completion_attempts:
                break
    count += 1
    

len(multiple_choice_questions)

# 生成的数据集
df_mcq = pd.DataFrame.from_records(multiple_choice_questions)
df_mcq.head(3)

df_mcq.to_csv('data/train_qa.csv', index=None)

# 原始数据集(存放维基百科文章,维基百科的链接,维基百科的标题)
ori_df = pd.DataFrame(ds)

ori_df.head(3)

ori_df.to_csv('data/ori_dataset.csv', index=None)

df_final = df_mcq.merge(ori_df, how='left', left_on='ori_wiki', right_on= 'text')

df_final = df_final[['question', 'text', 'url', 'title']]
df_final.head()

df_final.to_csv("data/retrive_dataset.csv", index=False)

df_final = pd.read_csv("data/retrive_dataset.csv") # 为原始数据增加
df_final.head(3)

# df_maq = pd.read_csv('data/ext-data/train_qa.csv')
# ori_df = pd.read_csv('data/ext-data/ori_dataset.csv')
# df_final = df_mcq.merge(ori_df, how='left', left_on='ori_wiki', right_on= 'text')
# df_final = df_final[['question', 'text', 'url', 'title']]
# df_final.to_csv("data/ext-data/retrive_dataset.csv", index=False)


ori_df = pd.read_csv("data/ext-data/all-data-gpt3.5/ori_dataset.csv")
print(f'ori_df: {len(ori_df)}')
df_final = pd.read_csv("data/ext-data/all-data-gpt3.5/retrive_dataset.csv") # 为原始数据增加
print(f'df_final: {len(df_final)}')
df_final.head(3)

neg= []
for row in tqdm(df_final.iterrows(), total=len(df_final)):
    row = row[1]
    url = row.url
    question = row.question
    candidates = ori_df[ori_df['url'] != url] # 取所有不属于url的候选集合
    candidates = candidates.sample(n=5) # 采样十个
    candidates['ori_url'] = url
    candidates['question'] = question
    neg.append(candidates)

neg = pd.concat(neg)
neg = neg.drop_duplicates()

neg.to_csv('data/ext-data/all-data-gpt3.5/neg_sample_5.csv', index=None)

len(neg)

neg = pd.read_csv('data/ext-data/all-data-gpt3.5/neg_sample_10.csv')
neg.head(3)

val_id = ori_df.url.unique()

len(val_id)

len(ori_df)

# ori_data = pd.read_csv("data/ori_dataset.csv")
# val_id = ori_data.url.unique()
# len(val_id)

# len(ori_data)

val_id = val_id[:100]
np.save('data/ext-data/all-data-gpt3.5/val_id_top100',val_id)

df = pd.read_csv('data/ext-data/all-data-gpt3.5/retrive_dataset.csv')  # 正样本

dev_ids = np.load('data/ext-data/all-data-gpt3.5/val_id_top100.npy', allow_pickle=True)
dev_df = df[df['url'].isin(dev_ids)]
len(dev_df)
